Cleanup usage statement in integration test
Note that none of the invocations of this function other than
in response to -h are ever invoked since the error() function
exits before usage() is called.  This is correct behavior, as
a usage statement should not be printed in response to an error,
and should never be written to stderr.