Retain exit status as first parameter
Note that I strongly object to this change.  A usage statement
should never be written to stderr, and emitting a usage statement
should not result in a non-zero exit status.  Since this functionality
is not currently used, IMO it ought to be discarded, but is retained
in this branch for backwards compatibility.