Remove check that BigQuery table exists for batch load
This check is part of the batch loading process. The check is made when files are being staged to GCS. At a certain scale, you begin to exceed BigQuery rate limits for API calls, documented at https://cloud.google.com/bigquery/quotas#api_requests.

The check could be wrapped and re-tried with backoff, or moved out of the critical path, but as the comment says the tables should be created at connector startup, I figured the best thing is to remove it and focus on making sure table creation at startup works.