MINOR: Handle new "table is deleted" error from BigQuery
Previously, if a table was deleted and recreated in a short time period, some writes to that table ran the risk of being silently dropped by BigQuery.

It seems like BigQuery handles that case now by returning a new error for writes to recently-recreated tables with the same 404 code and `notFound` reason, but a slightly different message (something like "Not found: table Table is deleted: <table>" (the "table Table" part is not a typo; this is what the responses actually look like)).

We can handle this new error the same way we handle missing table errors already. It may not be super useful for everyday users of the connector but it makes integration testing (and possibly upsert/delete logic) much easier since we can now wipe out tables left over from previous tests and then spin up connector instances that write to them immediately after.